+++ oauth2_provider/contrib/rest_framework/__init__.py
@@ -1,4 +1,6 @@
 # flake8: noqa
 from .authentication import OAuth2Authentication
-from .permissions import TokenHasScope, TokenHasReadWriteScope, TokenHasResourceScope
-from .permissions import IsAuthenticatedOrTokenHasScope
+from .permissions import (
+    TokenHasScope, TokenHasReadWriteScope, TokenHasMethodScopeAlternative,
+    TokenHasResourceScope, IsAuthenticatedOrTokenHasScope
+)